Move Products related tests to ProductsViewModelTest
